Clean up sample
